# full support for duplicate dependencies with overlapping markers