--- a/tests/installation/fixtures/with-multiple-updates.test
+++ b/tests/installation/fixtures/with-multiple-updates.test
@@ -9,8 +9,8 @@ files = []
 [package.dependencies]
 B = ">=1.0.1"
 C = [
-    {version = "^1.0", markers = "python_version >= \"2.7\" and python_version < \"2.8\""},
-    {version = "^2.0", markers = "python_version >= \"3.4\" and python_version < \"4.0\""},
+    {version = ">=1.0,<2.0", markers = "python_version >= \"2.7\" and python_version < \"2.8\""},
+    {version = ">=2.0,<3.0", markers = "python_version >= \"3.4\" and python_version < \"4.0\""},
 ]
 
 [[package]]