… tests/puzzle/test_solver.py