Resolves: #5066

Requires: python-poetry/poetry-core#530, python-poetry/poetry-core#546, python-poetry/poetry-core#547


- replace `_merge_dependencies_by_marker` and `_handle_any_marker_dependencies` by the more general `_resolve_overlapping_markers`
- raise `IncompatibleConstraintsError` if there are conflicts (due to incompatible version constraints or source types) in the constraints of a package

Tested with [opencv](https://pypi.org/project/opencv-python/4.6.0.66/), which has the most duplicate dependencies with overlapping markers I know about:
```
numpy (>=1.13.3) ; python_version < "3.7"
numpy (>=1.21.2) ; python_version >= "3.10"
numpy (>=1.21.2) ; python_version >= "3.6" and platform_system == "Darwin" and platform_machine == "arm64"
numpy (>=1.19.3) ; python_version >= "3.6" and platform_system == "Linux" and platform_machine == "aarch64"
numpy (>=1.14.5) ; python_version >= "3.7"
numpy (>=1.17.3) ; python_version >= "3.8"
numpy (>=1.19.3) ; python_version >= "3.9"
```

Especially the markers with `platform_system` and `platform_machine`, which overlap with most other markers make this example quite complex.

~~TODO: Investigation of the influence on the performance for (more) real-world examples~~



<!-- This is just a reminder about the most common mistakes. Please make sure that you tick all *appropriate* boxes.  But please read our [contribution guide](https://python-poetry.org/docs/contributing/) at least once, it will save you unnecessary review cycles! -->

- [x] Added **tests** for changed code.
- [ ] Updated **documentation** for changed code.

<!-- If you have *any* questions to *any* of the points above, just **submit and ask**!  This checklist is here to *help* you, not to deter you from contributing! -->