@radoering 2023-06-03
Times for `poetry lock` with a warm cache:

|`pyproject.toml` from ...| time without PR | time with PR |
|---|---|---|
|#4670|1 s|1 s|
|#4870|58 s|58 s|
|shootout example|16 s|10 s|
|large internal project|210 s|53 s|

Number of overrides:

|`pyproject.toml` from ...| number of overrides without PR | number of overrides with PR |
|---|---|---|
|#4670|24|24|
|#4870|52|52|
|shootout example|28|18|
|large internal project|93|15|

Considering the examples, the performance of our marker arithmetic seems to be fast enough to handle this PR: There is no measurable performance regression even in cases where the number of overrides does not change.