@dimbleby 2023-06-18
Here's an example in which `_resolve_overlapping_markers()` has to do too much work:
```toml
[tool.poetry.dependencies]
python = "^3.5"
oldest-supported-numpy = "^2022.11.19"
```
see https://github.com/scipy/oldest-supported-numpy/blob/main/setup.cfg