@radoering 2023-06-18
I can't get a solution (within a reasonable amount of time) with or without this PR. Thus, probably not a blocker?