@radoering 2023-06-19
After fixing the requirements of `oldest-supported-numpy` (cf scipy/oldest-supported-numpy#71) and with python-poetry/poetry-core#608, which is required to get sensible results for this package anyway, I tried some ranges with and without the PR:

Python constraint|without PR|with PR
---|---|---
^3.10|1 s|1 s
^3.9|44 s|13 s
\>=3.8,<3.10|> 1.000 s|69 s
\>=3.5,<3.8|858 s|318 s

Larger ranges still took too long. Maybe, I was just lucky but the results are pretty clear...