@dimbleby 2023-06-19
Interesting, do the changes you have made to oldest-supported-numpy have a negative effect on poetry performance?

without the PR and also without those changes I see:
| Python constraint | Time |
|-------------------|------|
| >=3.5,<3.8 | 5s |
| >=3.8,<3.10 | 3s |
| ^3.9 | 3s |
| ^3.10 | 1s |

where the first of those fails to find a solution and some of the others ought to (because the code without the PR is not correct); but nevertheless these times are _much_ faster than you are reporting.

I'm all in favour of concluding that the new code is both more correct and also faster than the old code, that would be lovely.  I'm just a bit confused about why we are seeing such different results.