@radoering 2023-06-20
The changes to oldest-supported-numpy and python-poetry/poetry-core#608 both have a negative impact on performance.