@dimbleby 2023-06-20
I haven't profiled, so this might be a bad or irrelevant idea...

I wonder whether `_is_relevant_marker()` is slow: this bit looks like quite a lot of work:
```python
            and self._python_constraint.allows_any(
                get_python_constraint_from_marker(marker)
            )
```
thought: the project python constraint could be transformed into an equivalent marker (once, up-front), and then intersected with `used_marker_intersection`.  If that intersection is (not) empty then `used_marker_intersection` is (not) irrelevant to the current project.

As I say, I haven't tested whether that's definitely a slow bit that is worth optimising: and anyway it isn't completely obvious that this would even be an improvement.  So don't block the merge on this idea!