@dimbleby 2023-06-20
and in fact that observation can be encouraged by putting the positive markers early on in the intersection, like this (could do with prettying up!):
```python
              markers = (
                  dep.marker if use else dep.marker.invert()
                  for use, dep in sorted(
                      zip(uses, dependencies), key=lambda ud: ud[0], reverse=True
                  )
              )
```
That's relying on the heuristic that if there are a lot of markers present, then surely they most don't intersect one another.  But that seems quite plausible, it would be quite hard to write them anyway