@radoering 2023-06-21
`_is_relevant_marker()` seems to be irrelevant. 😉

The other suggestion looks very promising. (I have to run some additional tests later.)