@dimbleby 2023-06-21
An easy win in `_merge_dependencies_by_constraint()`:
```diff
             for deps in by_constraint.values():
-                new_markers = [dep.marker for dep in deps]
                 dep = deps[0]
-                dep.marker = marker_union(*new_markers)
+                if len(deps) > 1:
+                    new_markers = (dep.marker for dep in deps)
+                    dep.marker = marker_union(*new_markers)
```
no need to perform all the dnf-and-cnf complexity when trying to merge a list that only contained a single marker in the first place