@radoering 2023-06-23
I applied both suggestions and checked that there is no negative impact on the examples in https://github.com/python-poetry/poetry/pull/7257#issuecomment-1575046996.

Measuring https://github.com/python-poetry/poetry/pull/7257#issuecomment-1597713699 again:

Python constraint|before optimization|after optimization 1|after optimization 2
---|---|---|---
^3.9|13 s|10 s|4 s
\>=3.8,<3.10|69 s|22 s|3 s
\>=3.5,<3.8|318 s|15 s|9 s

The only drawback of the second optimization is that the locked markers become more verbose. See `test_solver_duplicate_dependencies_with_overlapping_markers_complex`. (Of course, the performance improvement is more important than consise markers in the lock file.)