@dimbleby 2023-06-23
Seems like what we're seeing here is that the performance of marker operations is still expensive, when they get sufficiently complicated.  Nearly all of the time in the solving is now taken up with performing the single intersection of the final combination of markers (the case where all `uses` are `False`).  Improvements recorded above are great: but with broader python constraints this package is still too slow to solve.

`oldest-supported-numpy` is certainly more complicated than most, but it is a real found-in-the-wild example.  I don't think that it's necessary to completely solve that to merge this! Especially given the evidence that before this MR poetry was also nowhere near to finding a solution, so this isn't making things worse.

However, if you have any good ideas then this still would be worth addressing....

I suspect that we are doing a lot of repeated work to normalize the same markers again and again.  I don't think I'm going to get to trying anything out soon, but I wonder if there are wins to be had either by caching answers, or having markers remember that they are already in DNF (therefore another dnf() can be made a no-op).